spread into folders
