+++ internal/server/disaster.go
@@ -1,0 +1,9 @@
+package server
+
+// probeDisaster checks if a disaster occurs this turn
+func (s *SOMASServer) probeDisaster() (bool, error) {
+	s.logf("start probeDisaster")
+	defer s.logf("finish probeDisaster")
+	// TOOD:- env team
+	return false, nil
+}
