+++ internal/server/iifo.go
@@ -1,0 +1,10 @@
+package server
+
+// runIIFO : IIFO makes recommendations about the optimal (and fairest) contributions this term.
+// to mitigate common risk dilemma.
+func (s *SOMASServer) runIIFO() error {
+	s.logf("start runIIFO")
+	defer s.logf("finish runIIFO")
+	// TODO:- IIFO team
+	return nil
+}
