+++ internal/server/iigo.go
@@ -1,0 +1,9 @@
+package server
+
+// runIIGO : IIGO decides rule changes, elections, sanctions
+func (s *SOMASServer) runIIGO() error {
+	s.logf("start runIITO")
+	defer s.logf("finish runIITO")
+	// TOOD:- IIGO team
+	return nil
+}
