+++ internal/server/iito.go
@@ -1,0 +1,10 @@
+package server
+
+// runIITO : IITO makes recommendations about the optimal (and fairest) contributions this term
+// to mitigate the common pool dilemma
+func (s *SOMASServer) runIITO() error {
+	s.logf("start runIITO")
+	defer s.logf("finish runIITO")
+	// TOOD:- IITO team
+	return nil
+}
